+++ src/Service/Service.ts
@@ -3,13 +3,20 @@ import { FindOptions, Model, ModelCtor } from 'sequelize/types';
 import createError from 'http-errors';
 
 export class Service<Model extends Object = {}> {
-  private model: ModelCtor<any>;
+  private model?: ModelCtor<any>;
 
-  constructor(model: ModelCtor<any>) {
+  constructor(model?: ModelCtor<any>) {
     this.model = model;
   }
 
+  private noModelError = () => {
+    throw new Error(`Service can't be initialized, properly model is missing`);
+  };
+
   public count = async (query?: FindOptions) => {
+    if (!this.model) {
+      return this.noModelError();
+    }
     try {
       return await this.model.count(query);
     } catch (error) {
@@ -18,6 +25,9 @@ export class Service<Model extends Object = {}> {
   };
 
   public create = async (body: Model) => {
+    if (!this.model) {
+      return this.noModelError();
+    }
     try {
       return await this.model.create(body);
     } catch (error) {
@@ -26,6 +36,9 @@ export class Service<Model extends Object = {}> {
   };
 
   public deleteById = async (id: string | number) => {
+    if (!this.model) {
+      return this.noModelError();
+    }
     try {
       return await this.model.destroy({ where: { id } });
     } catch (error) {
@@ -34,6 +47,9 @@ export class Service<Model extends Object = {}> {
   };
 
   public find = async (query?: FindOptions) => {
+    if (!this.model) {
+      return this.noModelError();
+    }
     try {
       return await this.model.findAll(query);
     } catch (error) {
@@ -42,6 +58,9 @@ export class Service<Model extends Object = {}> {
   };
 
   public findById = async (id: number | string) => {
+    if (!this.model) {
+      return this.noModelError();
+    }
     try {
       return await this.model.findOne({ where: { id } });
     } catch (error) {
@@ -50,6 +69,9 @@ export class Service<Model extends Object = {}> {
   };
 
   public findOne = async (query?: FindOptions) => {
+    if (!this.model) {
+      return this.noModelError();
+    }
     try {
       return await this.model.findOne(query);
     } catch (error) {
@@ -58,6 +80,9 @@ export class Service<Model extends Object = {}> {
   };
 
   public updateById = async (id: string | number, body: Model) => {
+    if (!this.model) {
+      return this.noModelError();
+    }
     try {
       const data = await this.model.findOne({ where: { id } });
       if (data) {
